♻️ 8GE5QM acr: decompose generate helpers
