+++ .agentplane/tasks/202605290253-8GE5QM/README.md
@@ -1,0 +1,124 @@
+---
+id: "202605290253-8GE5QM"
+title: "ACR generate decomposition"
+status: "DOING"
+priority: "med"
+owner: "CODER"
+revision: 4
+origin:
+  system: "manual"
+depends_on: []
+tags:
+  - "acr"
+  - "code"
+  - "hotspot"
+verify: []
+plan_approval:
+  state: "approved"
+  updated_at: "2026-05-29T02:53:49.093Z"
+  updated_by: "ORCHESTRATOR"
+  note: null
+verification:
+  state: "pending"
+  updated_at: null
+  updated_by: null
+  note: null
+  attempts: 0
+commit: null
+comments:
+  -
+    author: "CODER"
+    body: "Start: decompose ACR generate helpers while preserving generated record behavior and digest validation."
+events:
+  -
+    type: "status"
+    at: "2026-05-29T02:54:05.771Z"
+    author: "CODER"
+    from: "TODO"
+    to: "DOING"
+    note: "Start: decompose ACR generate helpers while preserving generated record behavior and digest validation."
+doc_version: 3
+doc_updated_at: "2026-05-29T02:54:05.771Z"
+doc_updated_by: "CODER"
+description: "Decompose packages/agentplane/src/commands/acr/generate.ts by extracting ACR extension and residual-risk helpers while preserving generated Agent Change Record behavior."
+sections:
+  Summary: |-
+    ACR generate decomposition
+
+    Decompose packages/agentplane/src/commands/acr/generate.ts by extracting ACR extension and residual-risk helpers while preserving generated Agent Change Record behavior.
+  Scope: |-
+    - In scope: Decompose packages/agentplane/src/commands/acr/generate.ts by extracting ACR extension and residual-risk helpers while preserving generated Agent Change Record behavior.
+    - Out of scope: unrelated refactors not required for "ACR generate decomposition".
+  Plan: |-
+    Plan:
+    1. Start a branch_pr worktree for CODER.
+    2. Extract ACR blueprint/context extension projection and residual-risk/check helper logic from packages/agentplane/src/commands/acr/generate.ts into focused modules.
+    3. Preserve generateAcr public API, ACR schema validation, record digest behavior, and output path behavior.
+    4. Verify with ACR command tests, typecheck, arch/knip/lint/format, and hotspot threshold report.
+    5. Open PR, wait for hosted checks, merge to main, close lifecycle, and clean worktree.
+
+    Acceptance:
+    - Generated ACR records remain schema-valid and digest-compatible for existing tests.
+    - generate.ts drops below the 400-line hotspot warning threshold.
+    - Runtime hotspot warning count decreases from 21 to 20 without introducing new warning-sized runtime modules.
+    - No unrelated ACR validate, diff, remediation, or CLI parsing changes.
+  Verify Steps: |-
+    PLANNER fallback scaffold. Replace with task-specific acceptance checks when PLANNER context is available.
+
+    1. Review the changed artifact or behavior for the `code` task. Expected: the requested outcome is visible and matches the approved scope.
+    2. Run the most relevant validation step for the `code` task. Expected: it succeeds without unexpected regressions in touched scope.
+    3. Compare the final result against the task summary and scope. Expected: any remaining follow-up is explicit in ## Findings.
+  Verification: |-
+    <!-- BEGIN VERIFICATION RESULTS -->
+    <!-- END VERIFICATION RESULTS -->
+  Rollback Plan: |-
+    - Revert task-related commit(s).
+    - Re-run required checks to confirm rollback safety.
+  Findings: ""
+id_source: "generated"
+---
+## Summary
+
+ACR generate decomposition
+
+Decompose packages/agentplane/src/commands/acr/generate.ts by extracting ACR extension and residual-risk helpers while preserving generated Agent Change Record behavior.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/commands/acr/generate.ts by extracting ACR extension and residual-risk helpers while preserving generated Agent Change Record behavior.
+- Out of scope: unrelated refactors not required for "ACR generate decomposition".
+
+## Plan
+
+Plan:
+1. Start a branch_pr worktree for CODER.
+2. Extract ACR blueprint/context extension projection and residual-risk/check helper logic from packages/agentplane/src/commands/acr/generate.ts into focused modules.
+3. Preserve generateAcr public API, ACR schema validation, record digest behavior, and output path behavior.
+4. Verify with ACR command tests, typecheck, arch/knip/lint/format, and hotspot threshold report.
+5. Open PR, wait for hosted checks, merge to main, close lifecycle, and clean worktree.
+
+Acceptance:
+- Generated ACR records remain schema-valid and digest-compatible for existing tests.
+- generate.ts drops below the 400-line hotspot warning threshold.
+- Runtime hotspot warning count decreases from 21 to 20 without introducing new warning-sized runtime modules.
+- No unrelated ACR validate, diff, remediation, or CLI parsing changes.
+
+## Verify Steps
+
+PLANNER fallback scaffold. Replace with task-specific acceptance checks when PLANNER context is available.
+
+1. Review the changed artifact or behavior for the `code` task. Expected: the requested outcome is visible and matches the approved scope.
+2. Run the most relevant validation step for the `code` task. Expected: it succeeds without unexpected regressions in touched scope.
+3. Compare the final result against the task summary and scope. Expected: any remaining follow-up is explicit in ## Findings.
+
+## Verification
+
+<!-- BEGIN VERIFICATION RESULTS -->
+<!-- END VERIFICATION RESULTS -->
+
+## Rollback Plan
+
+- Revert task-related commit(s).
+- Re-run required checks to confirm rollback safety.
+
+## Findings
