+++ .agentplane/tasks/202605290253-8GE5QM/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605290253-8GE5QM`
+Title: ACR generate decomposition
+Canonical task record: `.agentplane/tasks/202605290253-8GE5QM/README.md`
+
+## Summary
+
+ACR generate decomposition
+
+Decompose packages/agentplane/src/commands/acr/generate.ts by extracting ACR extension and residual-risk helpers while preserving generated Agent Change Record behavior.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/commands/acr/generate.ts by extracting ACR extension and residual-risk helpers while preserving generated Agent Change Record behavior.
+- Out of scope: unrelated refactors not required for "ACR generate decomposition".
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T02:54:05.885Z
+- Branch: task/202605290253-8GE5QM/acr-generate-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
