+++ .agentplane/tasks/202605290253-8GE5QM/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 8GE5QM task: ACR generate decomposition [202605290253-8GE5QM]
