+++ .agentplane/tasks/202605290253-8GE5QM/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-29T02:54:05.885Z
+
+## Task
+
+- Task: `202605290253-8GE5QM`
+- Title: ACR generate decomposition
+- Status: DOING
+- Branch: `task/202605290253-8GE5QM/acr-generate-decomposition`
+- Canonical task record: `.agentplane/tasks/202605290253-8GE5QM/README.md`
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T02:54:05.885Z
+- Branch: task/202605290253-8GE5QM/acr-generate-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
